javascript/dealsumm-payment/django/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0

